Functional FCQ database
fcq database consists of individual fcq reports for a single class; i.e. each object in the fcq table represents an individual fcq report. Each report contains 2 foreign keys: one pointing to the class/course member that the fcq is for; the other points to a teacher object for the teacher that taught said course. The database is large (over 100k+ objects), but covers all fcq data from the past decade and does so effeciently throughout several database tables.