User page updates
Created a new user profile page that will act as a "hub" for the user; displaying all information and stats gathered about the user. The page will also feature an upload button that allows the user to upload a pdf of thier CU audit. This will be in the next commit. A seperate page was also created for the user to view the settings/info about their account. Here, they can change their password or any other information pertaining to their profile.