add upload ability to site
Users can now upload files to the site; this will be used to upload user audits. Current state of feature is FAR from finished; NO documents will be saved (contents will be read and file will immidiately be deleted), handeling function needs to be added, and entire presentation of feature needs to be redone.