Add file verification
Added file type verification to ensure only pdf's are uploaded. Also added "feature" of uploading files, handeling files, and deletion of files to ensure no files are stored on server disk.

NOTE: The process of temporarily storing the files did NOT use the built in temp-file handeling django offers. Instead, we save the file, interact with it, and autimatically delete the file. In the end, the same goal is accomplished.