UserAudit db implimentation
takes classes from audit and correctly enters the information into the database. If a user has taken a class that is not in the database, the class is added (no course title is given to the course however) to the class table in the database.

NOTE: Whenever a user uploads an audit pdf, all stored audit entries for the user will be deleted before adding the "new" audit information.